Re-added helping types: phone, halving & audience, created a test case for audience's help, restructured app layout to packages
