+++ millionaire/runner.py
@@ -1,0 +1,11 @@
+from millionaire.quiz import quiz
+
+
+
+def main():
+    # quiz.intro()
+    quiz.quiz()
+
+
+if __name__ == "__main__":
+    main()
