+++ save_us.py
@@ -6,7 +6,7 @@ def file_opening(filename, mode):
                 list_of_file.append(line)
     return list_of_file
 
-text=file_opening("data/text_files/phone_help_scenarios/yoda_master.txt", 'r')
+text=file_opening("data/text_files/yoda_master_phone.txt", 'r')
 if text[2][0] == 'I call the Force for help!':
     print(''.join(text[0]))
     print(good_answer+"\n")
